% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineSixMetrics.R
\name{CombineSixMetrics}
\alias{CombineSixMetrics}
\title{Combine six metrics into one summary}
\usage{
CombineSixMetrics(f_dat = ipmopat_exampprocdat, f_nsamp = 10000)
}
\arguments{
\item{f_dat}{Output from Ratings2Distributions function.}
}
\value{
A dataframe
}
\description{
Combine six metrics into one summary
}
