% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm_exampdat.R
\docType{data}
\name{ipm_exampdat}
\alias{ipm_exampdat}
\title{Example input data}
\format{
A data frame with 6 rows and 7 variables:
\describe{
\item{title}{A title describing the comparison being done}
\item{weight}{Weights given to each metric}
\item{short}{Weights given to each metric}
\item{ccp_rating}{Rating of CCP scenario}
\item{ccp_confidence}{Confidence in rating of CCP scenario}
\item{ipm_rating}{Rating of IPM scenario}
\item{ipm_confidence}{Confidence in rating of IPM scenario}
}
}
\usage{
ipm_exampdat
}
\description{
A dataset containing dummy ratings and confidences for the six metrics
}
\keyword{datasets}
