% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipm_exinput}
\alias{ipm_exinput}
\title{Example elicitation data (fair dust on lettuce)}
\format{
A data frame with an expanding number of rows and 4 variables:
\describe{
\item{title}{The user provided title for the comparison, should include organization, crop, and intervention}
\item{weight}{The weight assigned to each of the six impact categories}
\item{short}{The short way to refer to the impact category}
\item{ccp_rating}{The rating for the current commercial practice (CCP)
scenario for a given impact category
(very low impact, low impact, medium impact, high impact, very high impact)}
\item{ccp_confidence}{The confidence the user has in the given rating (L, M, H, VH)}
\item{ipm_rating}{The rating for the new integrated pest management (IPM) scenario for a given impact category}
\item{ipm_confidence}{The confidence the user has in the given rating}
}
}
\usage{
ipm_exinput
}
\description{
An example of the elicitation data format to be input into the package function
}
\keyword{datasets}
