% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipm_impactcatsinfo}
\alias{ipm_impactcatsinfo}
\title{More detailed description of the six impact categories}
\format{
A data frame with 6 rows and 3 variables:
\describe{
\item{short}{The short notation to refer to the impact category}
\item{value_metric}{A longer description of the impact category}
\item{question}{The question used in the elicitation paperwork (more or less)}
}
}
\usage{
ipm_impactcatsinfo
}
\description{
Includes the question used in the elicitation form, the short description, and the long description
}
\keyword{datasets}
