% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeImage4R.R
\name{MakeImage4R}
\alias{MakeImage4R}
\title{Intakes elicitation data}
\usage{
MakeImage4R(f_dat = ipm_exinput)
}
\arguments{
\item{f_dat}{The elicited data, default is to use ipm_exinput}
}
\value{
An image
}
\description{
Intakes elicitation data
}
