% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonsterFxn.R
\name{MonsterFxn}
\alias{MonsterFxn}
\title{Intakes elicitation data}
\usage{
MonsterFxn(f_dat = ipm_exinput)
}
\arguments{
\item{f_dat}{The elicited data.}
}
\value{
An image
}
\description{
Intakes elicitation data
}
