% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ratings2Distributions.R
\name{Ratings_2_Distributions}
\alias{Ratings_2_Distributions}
\title{Create beta distributions for each rating and confidence pair}
\usage{
Ratings_2_Distributions(f_dat = impopat_exampdat)
}
\arguments{
\item{f_dat}{The elicited data.}
}
\value{
A dataframe
}
\description{
Create beta distributions for each rating and confidence pair
}
