% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeResultsBarFig.R
\name{VisualizeResultsBarFig}
\alias{VisualizeResultsBarFig}
\title{Combine six metrics into one summary}
\usage{
VisualizeResultsBarFig(f_dat1 = ipm_exampprocdat, f_dat2 = ipm_exampcombdat)
}
\arguments{
\item{f_dat1}{Output from Ratings2Distributions function.}

\item{f_dat2}{Output from CombineSixMetrics function.}
}
\value{
A figure using dot sizes to represent value
}
\description{
Combine six metrics into one summary
}
